refactor(context): split docker context types
